Refactor: mock 데이터 파일 이동
