+++ src/apis/artBuyerPage/type.ts
@@ -1,8 +1,12 @@
-import { TArtist } from '@/types/artist';
+/**
+ * 구매한 작품 데이터 타입
+ * @author 노찬영
+ **/
 
-export type TGetArtistListResponse = TArtist[];
-
-export type TGetArtistListRequestParams = {
-  page?: number;
-  keyword?: string;
+export type TPurchasedArtworkResponse = {
+  artwork_id: number;
+  author_name: string;
+  title: string;
+  size: string;
+  is_liked: boolean;
 };
