+++ src/apis/type.ts
@@ -1,5 +1,6 @@
 //공통 responseDTO
 export type TGetResponse<T> = {
-  success: boolean;
+  isSuccess: boolean;
   response: T;
+  result: T;
 };
