+++ src/pages/artBuyerPage/components/menuMyPage/Auction/index.tsx
@@ -10,7 +10,7 @@ import {
 
 import { ArtBuyerDataProps } from '@/types/artBuyer';
 
-import { artBuyerData as rawArtBuyerData } from '@/pages/artBuyerPage/constants/artBuyer';
+import { artBuyerData as rawArtBuyerData } from '@/constants/mocks';
 
 /**
  * @description 작품 구매자의 경매 내역을 표시하는 컴포넌트
