+++ src/pages/authorPage/components/managementWorks/auctioningWorks/index.tsx
@@ -10,9 +10,9 @@ import {
 
 import { ArtworksExhibitionsDataProps } from '@/types/author';
 
-import { artworksExhibitionsData as rawArtworksExhibitionData } from '@/pages/authorPage/constants/artworksExhibitions';
+import { artworksExhibitionsData as rawArtworksExhibitionData } from '@/constants/mocks';
 
-// 명시적으로 ArtBuyerDataProps 타입 지정
+// 명시적으로 ArtworksExhibitionsDataProps 타입 지정
 const artworksExhibitionsData: ArtworksExhibitionsDataProps =
   rawArtworksExhibitionData;
 const { auction_artworks } = artworksExhibitionsData.result;
