+++ src/pages/authorPage/components/managingProfiles/allInfo/index.tsx
@@ -15,9 +15,9 @@ import {
 
 import { AuthorDataProps } from '@/types/author';
 
-import { authorData as rawAuthorData } from '@/pages/authorPage/constants/author';
+import { authorData as rawAuthorData } from '@/constants/mocks';
 
-// 명시적으로 ArtBuyerDataProps 타입 지정
+// 명시적으로 AuthorDataProps 타입 지정
 const authorData: AuthorDataProps = rawAuthorData;
 const {
   author_name,
